requirements.txt
python-sql
shutils
gitpython
pypi-json